// src/app/models/full-sorteio.ts
export class FullSorteio {
  id: number;
  descricao: string;
  ganhador: string;
  imagem: string | ArrayBuffer;
  numeroSorteado: string;
  status: number;
  statusDescricao: string;
  titulo: string;
	rifinha: boolean;
  idSorteioRifinha: number;
  nomeSorteioRifinha: string;
	ocultar: boolean;
  valor: number;
  exibirData: boolean;
  whatsapp: string;
  keywords: string;
  adicional1: string;
  adicional2: string;
  texto = '';
  dataSorteio: string;
  numeroInicial: number;
  numeroFinal: number;
}